% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{covsum}
\alias{covsum}
\title{Get covariate summary dataframe}
\usage{
covsum(
  data,
  covs,
  maincov = NULL,
  digits = 1,
  numobs = NULL,
  markup = TRUE,
  sanitize = TRUE,
  nicenames = TRUE,
  IQR = FALSE,
  digits.cat = 0,
  testcont = c("rank-sum test", "ANOVA"),
  testcat = c("Chi-squared", "Fisher")
)
}
\arguments{
\item{data}{dataframe containing data}

\item{covs}{character vector with the names of columns to include in table}

\item{maincov}{covariate to stratify table by}

\item{digits}{number of digits for summarizing mean data}

\item{numobs}{named list overriding the number of people you expect to have the covariate}

\item{markup}{boolean indicating if you want latex markup}

\item{sanitize}{boolean indicating if you want to sanitize all strings to not break LaTeX}

\item{nicenames}{booling indicating if you want to replace . and _ in strings with a space}

\item{IQR}{boolean indicating if you want to display the inter quantile range (Q1,Q3) as opposed to (min,max) in the summary for continuous variables}

\item{digits.cat}{number of digits for the proportions when summarizing categorical data (default: 0)}

\item{testcont}{test of choice for continuous variables,one of \emph{rank-sum} (default) or \emph{ANOVA}}

\item{testcat}{test of choice for categorical variables,one of \emph{Chi-squared} (default) or \emph{Fisher}}
}
\description{
Returns a dataframe corresponding to a descriptive table
}
\seealso{
\code{\link{fisher.test}},\code{\link{chisq.test}},\code{\link{wilcox.test}},\code{\link{kruskal.test}},and \code{\link{anova}}
}
\keyword{dataframe}
