% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ggkmcif_paste}
\alias{ggkmcif_paste}
\title{Plot KM and CIF curves with ggplot}
\usage{
ggkmcif_paste(list_gg)
}
\arguments{
\item{list_gg}{list containing the results of ggkmcif}
}
\description{
This function puts together a survival curve, and a number at risk table
}
