% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ggsurv}
\alias{ggsurv}
\title{Plot KM and CI curves using ggplot2}
\usage{
ggsurv(response, cov = NULL, data, type = NULL, times = NULL,
  table = TRUE, returns = FALSE, xlabs = "Time", ylabs = NULL,
  main = NULL, ystratalabs = NULL, ystrataname = nicename(cov),
  censor.marks = FALSE, HR = FALSE, HR.pval = FALSE,
  conf.type = "none", fsize = 15, nsize = 4.5, lsize = 1.5,
  psize = 4.5, ylims = c(0, 1), xlims = NULL, col = NULL,
  linetype = NULL, legend.pos = NULL, pval = TRUE, pval.pos = NULL,
  plot.event = 1, event = c("col", "linetype"), flip.CIF = FALSE,
  cut = NULL, ticklabs = NULL, logse = FALSE, eventlabs = NULL,
  event.name = NULL, ...)
}
\arguments{
\item{response}{covariate to stratify table by}

\item{cov}{character vector with the names of columns to include in table}

\item{data}{dataframe containing data}

\item{type}{...}

\item{times}{...}

\item{table}{boolean indicating if you want latex markup}

\item{returns}{boolean indicating if you want to sanitize all strings to not break LaTeX}

\item{xlabs}{booling indicating if you want to replace . and _ in strings with a space}

\item{ylabs}{boolean indicating if you want to display the inter quantile range (Q1,Q3) as opposed to (min,max) in the summary for continuous variables}

\item{main}{number of digits for the proportions when summarizing categorical data (default: 0)}

\item{ystratalabs}{test of choice for continuous variables, one of \emph{rank-sum} (default) or \emph{ANOVA}}

\item{ystrataname}{test of choice for categorical variables, one of \emph{Chi-squared} (default) or \emph{Fisher}}

\item{censor.marks}{...}

\item{HR}{...}

\item{HR.pval}{...}

\item{conf.type}{...}

\item{fsize}{...}

\item{nsize}{...}

\item{lsize}{...}

\item{psize}{...}

\item{ylims}{...}

\item{xlims}{......}

\item{col}{...}

\item{linetype}{...}

\item{legend.pos}{...}

\item{pval}{...}

\item{pval.pos}{...}

\item{plot.event}{...}

\item{event}{...}

\item{flip.CIF}{...}

\item{cut}{...}

\item{ticklabs}{...}

\item{logse}{...}

\item{eventlabs}{...}

\item{event.name}{...}

\item{...}{additional options passed to function}
}
\description{
Generates Kaplan-Meier and Cumulative Incidence plots including a numbers at risk table via ggplot2 with high level of customization.
}
\examples{
require(survival)
lung$sex<-factor(lung$sex)
ggsurv(c("time","status"), data=lung)
ggsurv(c("time","status"),"sex", data=lung)
}
\seealso{
\code{\link{fisher.test}}
}
\keyword{dataframe}
