% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{importCodes}
\alias{importCodes}
\title{Return a data frame of codes}
\usage{
importCodes(labelStr, delim = ",", codeName = "code", lblName = "label")
}
\arguments{
\item{labelStr}{in the format code1=label1,code2=label2}

\item{delim}{delimeter separating codes in labelStr, defaults to ','}

\item{codeName}{column name for codes, defaults to code}

\item{lblName}{column name for labels, defaults to label}
}
\description{
Accepts a string input in the form "code1=label1,code2=label2,.." and 
returns a data frame with a column of codes and a column of labels
}
