% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{modify_ggkmcif}
\alias{modify_ggkmcif}
\title{Plot KM and CIF curves with ggplot}
\usage{
modify_ggkmcif(list_gg)
}
\arguments{
\item{list_gg}{list containing the results of ggkmcif}
}
\description{
This function put together a survival curve, and a number at risk table
}
