% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{nestTable}
\alias{nestTable}
\title{Combine two table columns into a single column with levels of one nested within levels of the other.}
\usage{
nestTable(
  data,
  head_col,
  to_col,
  caption = NULL,
  indent = TRUE,
  boldheaders = T,
  hdr_prefix = "",
  hdr_suffix = "",
  tblOnly = FALSE
)
}
\arguments{
\item{data}{dataframe}

\item{head_col}{character value specifying the column name with the headers}

\item{to_col}{character value specifying the column name to add the headers into}

\item{caption}{table caption}

\item{indent}{Boolean should the original values in the to_col be indented}

\item{boldheaders}{Boolean should the header column values be bolded}

\item{hdr_prefix}{character value that will prefix headers}

\item{hdr_suffix}{character value that will suffix headers}

\item{tblOnly}{boolean indicating if the table should be formatted for printing or returned as a data frame}
}
\description{
This function accepts a data frame (via the data argument) and combines two columns into
a single column with values from the head_col serving as headers and values of the to_col displayed
underneath each header. The resulting table is then passed to outTable for printing and output, to use the 
grouped table as a data frame specify tblOnly=TRUE.
By default the headers will be bolded and the remaining columns indented.
}
