% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pcovsum}
\alias{pcovsum}
\title{Print covariate summary Latex}
\usage{
pcovsum(data, covs, maincov = NULL, TeX = FALSE, ...)
}
\arguments{
\item{data}{dataframe containing data}

\item{covs}{character vector with the names of columns to include in table}

\item{maincov}{covariate to stratify table by}

\item{TeX}{boolean indicating if you want to be able to view extra long tables in the LaTeX pdf. If TeX is T then the table will not convert properly to docx}

\item{...}{additional options passed to function  \code{\link{covsum}}}
}
\description{
Returns a dataframe corresponding to a descriptive table
}
\keyword{print}
