% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{petsum}
\alias{petsum}
\title{Print LaTeX event time summary}
\usage{
petsum(data, response, group = 1, times = c(12, 14), units = "months")
}
\arguments{
\item{data}{dataframe containing data}

\item{response}{character vector with names of columns to use for response}

\item{group}{string specifying the column name of stratification variable}

\item{times}{numeric vector of times you want survival time probabilities for.}

\item{units}{string indicating the unit of time. Use lower case and plural.}
}
\description{
Wrapper for the etsum function that prints paragraphs of text in LaTeX
}
\examples{
petsum(pembrolizumab,c('os_time','os_status'),"sex")
petsum(pembrolizumab,c('os_time','os_status'))
petsum(pembrolizumab,c('os_time','os_status'),"sex",c(1,2,3),"months")
}
\keyword{print}
