% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_univariate}
\alias{plot_univariate}
\title{Plot univariate relationships all on one plot
Need a hierarchy of how to display plots sensibly
If response is continuous
  For a numeric predictor -> scatterplot
  For a categorical predictor -> boxplot
If response is a factor
  For a numeric predictor -> boxplot
  For a categorical predictor -> barplot}
\usage{
plot_univariate(
  response,
  covs,
  data,
  showN = FALSE,
  na.rm = TRUE,
  response_title = NULL
)
}
\arguments{
\item{response}{character vector with names of columns to use for response}

\item{covs}{character vector with names of columns to use for covariates}

\item{data}{dataframe containing your data}

\item{showN}{boolean indicating whether sample sizes should be shown on the plots}

\item{na.rm}{boolean indicating whether na values should be shown or removed}

\item{response_title}{character value with title of the plot}
}
\description{
Plot univariate relationships all on one plot
Need a hierarchy of how to display plots sensibly
If response is continuous
  For a numeric predictor -> scatterplot
  For a categorical predictor -> boxplot
If response is a factor
  For a numeric predictor -> boxplot
  For a categorical predictor -> barplot
}
\keyword{plot}
