% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotuv}
\alias{plotuv}
\title{Plot bivariate relationships in a single plot}
\usage{
plotuv(
  response,
  covs,
  data,
  showN = FALSE,
  na.rm = TRUE,
  response_title = NULL
)
}
\arguments{
\item{response}{character vector with names of columns to use for response}

\item{covs}{character vector with names of columns to use for covariates}

\item{data}{dataframe containing your data}

\item{showN}{boolean indicating whether sample sizes should be shown on the plots}

\item{na.rm}{boolean indicating whether na values should be shown or removed}

\item{response_title}{character value with title of the plot}
}
\description{
This function is designed to accompany \code{\link{uvsum}} as a means
of visualising the results. It is not designed for publication, but rather
to facilitate explanations of the uvsum output.
}
\details{
The following hierarchy of how to display plots is used:
If response is continuous
  For a numeric predictor -> scatterplot
  For a categorical predictor -> boxplot
If response is a factor
  For a numeric predictor -> boxplot
  For a categorical predictor -> barplot
}
\keyword{plot}
