% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rm_covsum}
\alias{rm_covsum}
\title{Outputs a descriptive covariate table}
\usage{
rm_covsum(
  data,
  covs,
  maincov = NULL,
  caption = NULL,
  tableOnly = FALSE,
  covTitle = "Covariate",
  chunk_label,
  ...
)
}
\arguments{
\item{data}{dataframe containing data}

\item{covs}{character vector with the names of columns to include in table}

\item{maincov}{covariate to stratify table by}

\item{caption}{character containing table caption. If caption = NULL then 
notes about unstable estimates and p-value adjustments will be added to the
caption. To suppress the caption completely set caption='none'.}

\item{tableOnly}{Logical, if TRUE then a dataframe is returned, otherwise a 
formatted printed object is returned (default).}

\item{covTitle}{character with the names of the covariate column}

\item{chunk_label}{only used if output is to Word to allow cross-referencing}

\item{...}{additional options passed to function  \code{\link{covsum}}}
}
\value{
A formatted table displaying a summary of the covariates stratified by maincov
}
\description{
This is a wrapper function around covsum for use in Rmarkdown documents
}
\details{
Comparisons for categorical variables default to chi-square tests, but if there are counts of <5 then the Fisher Exact 
test will be used and if this is unsuccessful then a second attempt will be made computing p-values using MC simulation. 
If testcont='ANOVA' then the t-test with unequal variance will be used for two groups and an ANOVA will be used for three or more.
The statistical test used can be displayed by specifying show.tests=TRUE.
}
\seealso{
\code{\link{fisher.test}}, \code{\link{chisq.test}}, \code{\link{wilcox.test}}, \code{\link{kruskal.test}}, and \code{\link{anova}}
}
\keyword{dataframe}
