% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rm_ordsum}
\alias{rm_ordsum}
\title{Fit and format an ordinal logistic regression using polr from the {MASS} package. The parallel regression assumption can
be tested using the Brant test in the Brant package and visually. Only logistic ordinal regression is supported currently.}
\usage{
rm_ordsum(
  data,
  covs,
  response,
  reflevel,
  caption = NULL,
  showN = T,
  digits = 2,
  CIwidth = 0.95,
  excludeLevels = NULL,
  chunk_label
)
}
\arguments{
\item{data}{dataframe containing data [REQUIRED]}

\item{covs}{character vector with the names of columns to include in table [REQUIRED]}

\item{response}{ordinal outcome variable [REQUIRED]}

\item{reflevel}{manual specification of the reference level, must match level exactly}

\item{caption}{Table caption}

\item{showN}{logical, should sample sizes be shown for each lvel, defaults to TRUE}

\item{digits}{number of digits to display, defaults to}

\item{CIwidth}{level of significance for computing the confidence intervals, default is 0.95}

\item{excludeLevels}{a named list of levels to exclude from the response variable}

\item{chunk_label}{only used if output is to Word to allow cross-referencing}
}
\value{
A formatted table displaying the odds ratio associated with each covariate
}
\description{
Fit and format an ordinal logistic regression using polr from the {MASS} package. The parallel regression assumption can
be tested using the Brant test in the Brant package and visually. Only logistic ordinal regression is supported currently.
}
\keyword{Brant}
\keyword{ordinal}
\keyword{regression,}
\keyword{test}
