% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rm_uv_mv}
\alias{rm_uv_mv}
\title{Combine univariate and multivariable regression tables}
\usage{
rm_uv_mv(uvsumTable, mvsumTable, caption = NULL, tableOnly = F, chunk_label)
}
\arguments{
\item{uvsumTable}{Output from rm_uvsum, with tableOnly=T}

\item{mvsumTable}{Output from rm_mvsum. with tableOnly=T}

\item{caption}{table caption}

\item{tableOnly}{boolean indicating if unformatted table should be returned}

\item{chunk_label}{only used if output is to Word to allow cross-referencing}

\item{id}{character vector which identifies clusters. Only used for geeglm}
}
\description{
This function will combine rm_uvsum and rm_mvsum outputs into a single table
The tableOnly argument must be set to TRUE when tables to be combined are
created. The resulting table will be in the same order as the uvsum table
and will contain the same columns as the uvsum and mvsum tables, but the 
p-values will be combined into a single column. There must be variable over-
lap between the uvsum and mvsum tables and all variables in the mvsum table
must also appear in the uvsum table.
}
