% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{nestTable}
\alias{nestTable}
\title{Combine two table columns into a single column with levels of one nested within levels of the other.}
\usage{
nestTable(
  data,
  head_col,
  to_col,
  caption = NULL,
  indent = TRUE,
  boldheaders = TRUE,
  hdr_prefix = "",
  hdr_suffix = "",
  digits = 2,
  tableOnly = FALSE
)
}
\arguments{
\item{data}{dataframe}

\item{head_col}{character value specifying the column name with the headers}

\item{to_col}{character value specifying the column name to add the headers into}

\item{caption}{table caption}

\item{indent}{Boolean should the original values in the to_col be indented}

\item{boldheaders}{Boolean should the header column values be bolded}

\item{hdr_prefix}{character value that will prefix headers}

\item{hdr_suffix}{character value that will suffix headers}

\item{digits}{number of digits to round numeric columns to, wither a single number or a vector corresponding to the number of numeric columns}

\item{tableOnly}{boolean indicating if the table should be formatted for printing or returned as a data frame}
}
\description{
This function accepts a data frame (via the data argument) and combines two columns into
a single column with values from the head_col serving as headers and values of the to_col displayed
underneath each header. The resulting table is then passed to outTable for printing and output, to use the 
grouped table as a data frame specify tableOnly=TRUE.
By default the headers will be bolded and the remaining values indented. 
Note that it is possible to combine multiple tables (more than two) with this function.
}
\examples{
## Create models to predict mileage and horsepower and show the outputs together
# fit1 <- lm(mpg~cyl+disp+wt,data=mtcars)
# m1 <- rm_mvsum(fit1,tableOnly=T)
# m1$Response = 'mileage'
# fit2 <- lm(hp~cyl+disp+wt,data=mtcars)
# m2 <- rm_mvsum(fit2,tableOnly=T)
# m2$Response = 'horsepower'
# rbind(m1,m2)
# nestTable(rbind(m1,m2),head_col='Response',to_col='Covariate')
}
