% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{outTable}
\alias{outTable}
\title{Print tables to PDF/Latex HTML or Word}
\usage{
outTable(
  tab,
  to_indent = numeric(0),
  to_bold = numeric(0),
  caption = NULL,
  digits,
  align,
  chunk_label
)
}
\arguments{
\item{tab}{a table to format}

\item{to_indent}{numeric vector the length of nrow(tab) indicating which rows to indent}

\item{to_bold}{numeric vector the length of nrow(tab) indicating which rows to bold}

\item{caption}{table caption}

\item{digits}{number of digits to round numeric columns to, wither a single number or a vector corresponding to the number of numeric columns}

\item{align}{string specifying column alignment, defaults to left alignment of the first column and right alignment of all other columns}

\item{chunk_label}{only used knitting to Word docs to allow cross-referencing}
}
\description{
Output the table nicely to whatever format is appropriate. 
This is the output function used by the rm_* printing functions.
}
