% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pmvsum}
\alias{pmvsum}
\title{Print multivariate summary LaTeX table}
\usage{
pmvsum(model, data)
}
\arguments{
\item{model}{fitted model object}

\item{data}{dataframe containing data}
}
\description{
Returns a LaTeX table of the multivariate summary.
}
\keyword{print}
