% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rm_covsum}
\alias{rm_covsum}
\title{Returns a dataframe corresponding to a descriptive table for printing in Rmarkdown
The default output is a kable table for use in pdfs or html, but pander tables can be produced
for Word documents by specifying options('doc_type'='doc') in the setup chunk of the markdown document.}
\usage{
rm_covsum(
  data,
  covs,
  maincov = NULL,
  caption = NULL,
  excludeLevels = NULL,
  pvalue = TRUE,
  IQR = FALSE,
  tableOnly = FALSE,
  covTitle = "Covariate",
  chunk_label,
  ...
)
}
\arguments{
\item{data}{dataframe containing data}

\item{covs}{character vector with the names of columns to include in table}

\item{maincov}{covariate to stratify table by}

\item{caption}{character containing table caption}

\item{excludeLevels}{a named list of levels to exclude in the form list(varname =c('level1','level2')) these levels will be excluded from association tests}

\item{pvalue}{boolean indicating if p values should be displayed (may only want corrected p-values)}

\item{IQR}{boolean indicating if you want to display the inter quantile range (Q1,Q3) as opposed to (min,max) in the summary for continuous variables}

\item{tableOnly}{should a dataframe or a formatted print object be returned}

\item{covTitle}{character with the names of the covariate column}

\item{chunk_label}{only used if options("doc_type"='doc') to allow cross-referencing}

\item{...}{other variables passed to covsum and the table output function}
}
\value{
A formatted table displaying a summary of the covariates stratified by maincov
}
\description{
Returns a dataframe corresponding to a descriptive table for printing in Rmarkdown
The default output is a kable table for use in pdfs or html, but pander tables can be produced
for Word documents by specifying options('doc_type'='doc') in the setup chunk of the markdown document.
}
\seealso{
\code{\link{fisher.test}}, \code{\link{chisq.test}}, \code{\link{wilcox.test}}, \code{\link{kruskal.test}}, and \code{\link{anova}}
}
\keyword{dataframe}
