% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rm_covsum}
\alias{rm_covsum}
\title{Returns a dataframe corresponding to a descriptive table for printing in Rmarkdown
The default output is a kable table for use in pdfs or html, but pander tables can be produced
for Word documents by specifying options('doc_type'='doc') in the setup chunk of the markdown document.}
\usage{
rm_covsum(
  data,
  covs,
  maincov = NULL,
  caption = NULL,
  tableOnly = FALSE,
  covTitle = "Covariate",
  chunk_label,
  ...
)
}
\arguments{
\item{data}{dataframe containing data}

\item{covs}{character vector with the names of columns to include in table}

\item{maincov}{covariate to stratify table by}

\item{caption}{character containing table caption}

\item{tableOnly}{should a dataframe or a formatted print object be returned}

\item{covTitle}{character with the names of the covariate column}

\item{chunk_label}{only used if output is to Word to allow cross-referencing}

\item{...}{additional options passed to function  \code{\link{covsum}}}
}
\value{
A formatted table displaying a summary of the covariates stratified by maincov
}
\description{
Returns a dataframe corresponding to a descriptive table for printing in Rmarkdown
The default output is a kable table for use in pdfs or html, but pander tables can be produced
for Word documents by specifying options('doc_type'='doc') in the setup chunk of the markdown document.
}
\seealso{
\code{\link{fisher.test}}, \code{\link{chisq.test}}, \code{\link{wilcox.test}}, \code{\link{kruskal.test}}, and \code{\link{anova}}
}
\keyword{dataframe}
