% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rm_mvsum}
\alias{rm_mvsum}
\title{Format a regression model nicely for Rmarkdown}
\usage{
rm_mvsum(
  model,
  data,
  showN = FALSE,
  CIwidth = 0.95,
  caption = NULL,
  tableOnly = FALSE,
  p.adjust = "none",
  chunk_label
)
}
\arguments{
\item{model}{model fit}

\item{data}{data that model was fit on (an attempt will be made to extract
this from the model)}

\item{showN}{boolean indicating sample sizes should be shown for each comparison, can be useful for interactions}

\item{CIwidth}{width for confidence intervals, defaults to 0.95}

\item{caption}{table caption}

\item{tableOnly}{boolean indicating if unformatted table should be returned}

\item{p.adjust}{p-adjustments to be performed (Global p-values only)}

\item{chunk_label}{only used if output is to Word to allow cross-referencing}
}
\description{
This is a wrapper around mvsum for use with Rmarkdown
}
