% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rm_mvsum}
\alias{rm_mvsum}
\title{Output a multivariable model nicely in Rmarkdown
The default output is a kable table for use in pdfs or html, but pander tables can be produced
for Word documents by specifying options('doc_type'='doc') in the setup chunk of the markdown document.}
\usage{
rm_mvsum(
  model,
  data,
  showN = FALSE,
  CIwidth = 0.95,
  caption = NULL,
  tableOnly = FALSE,
  HolmGlobalp = FALSE,
  chunk_label
)
}
\arguments{
\item{model}{model fit}

\item{data}{data that model was fit on}

\item{showN}{boolean indicating sample sizes should be shown for each comparison, can be useful for interactions}

\item{CIwidth}{width for confidence intervals, defaults to 0.95}

\item{caption}{table caption}

\item{tableOnly}{boolean indicating if unformatted table should be returned}

\item{HolmGlobalp}{boolean indicting if a Holm-corrected p-value should be presented}

\item{chunk_label}{only used if output is to Word to allow cross-referencing}
}
\description{
Output a multivariable model nicely in Rmarkdown
The default output is a kable table for use in pdfs or html, but pander tables can be produced
for Word documents by specifying options('doc_type'='doc') in the setup chunk of the markdown document.
}
