% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rm_uvsum}
\alias{rm_uvsum}
\title{Output several univariate models nicely in a single table}
\usage{
rm_uvsum(
  response,
  covs,
  data,
  caption = NULL,
  tableOnly = FALSE,
  removeInf = T,
  HolmGlobalp = FALSE,
  chunk_label,
  ...
)
}
\arguments{
\item{response}{string vector with name of response}

\item{covs}{character vector with the names of columns to fit univariate models to}

\item{data}{dataframe containing data}

\item{caption}{table caption}

\item{tableOnly}{boolean indicating if unformatted table should be returned}

\item{removeInf}{boolean indicating if infinite estimates should be removed from the table}

\item{HolmGlobalp}{boolean indicting if a Holm-corrected p-value should be presented}

\item{chunk_label}{only used if output is to Word to allow cross-referencing}

\item{...}{additional options passed to function  \code{\link{uvsum}}}
}
\description{
Wrapper for the uvsum function for use with Rmarkdown.
}
